Refactored the article model to use uppercase in the filename and in the db.Article calls. Also rearranged file structure in handlebars to use partials.
