+++ server.js
@@ -33,7 +33,7 @@ mongoose.connect(MONGODB_URI);
 // GET all the articles that is set to false
 // and render them to the index.handlebars page
 app.get("/", function (req, res) {
-  db.article.find({ article: false })
+  db.Article.find({ article: false })
     .then(function (data) {
       // res.json(data);
       res.render("index", { article: data });
@@ -44,8 +44,8 @@ app.get("/", function (req, res) {
 
 // GET all the articles that is set to true
 // and render them to the article.handlebars page
-app.get("./article", function (req, res) {
-  db.article.find({ favorite: true })
+app.get("/article", function (req, res) {
+  db.Article.find({ favorite: true })
     .then(function (data) {
       // res.json(data);
       res.render("article", { article: data });
@@ -56,7 +56,7 @@ app.get("./article", function (req, res) {
 
 // POST an Article to the mongo database
 app.post("/api/index", function (req, res) {
-  db.index.create(req.body)
+  db.Article.create(req.body)
     .then(function () {
       // res.json(dbIndex);
       res.redirect("/")
@@ -69,7 +69,7 @@ app.post("/api/index", function (req, res) {
 // Will set the articles to whatever the value 
 // of the req.body.article boolean is
 app.put("/api/article/:id", function (req, res) {
-  db.articles.findByIdAndUpdate(req.params.id, { favorite: req.body.article }, { new: true })
+  db.Article.findByIdAndUpdate(req.params.id, { favorite: req.body.article }, { new: true })
     .then(function (dbArticle) {
       res.json(dbArticle);
     }).catch(function (err) {
